--- source/scenes/Registration/FormContext.js
@@ -1,11 +1,0 @@
-import { createContext } from 'react';
-
-export type RegistrationData = {
-  major: string,
-  graduationYear: string,
-  phoneNumber: string,
-};
-
-const FormContext = createContext();
-
-export default FormContext;
