+++ source/scenes/Registration/Panes/TeamInfo.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 
 import TextField from 'components/TextField';
-import FormContext from '../FormContext';
+import FormContext from 'components/FormContext';
 
 const TeamInfo = () => (
   <FormContext.Consumer>
